sugarcare_apis/requirements.txt
# Essential Dependencies for SugarCare Model Training
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic_core==2.14.1
python-multipart==0.0.6
requests==2.32.5
pillow==12.0.0
opencv-python==4.12.0.88
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.2
matplotlib==3.10.7
seaborn==0.13.2
joblib==1.5.2
torch==2.9.0
torchvision==0.24.0
python-dotenv==1.1.1
colorlog==6.10.1
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.1.1
argon2-cffi==23.1.0
reportlab==4.0.9
firebase-admin
google-auth>=2.0.0
